fix: correct import paths after directory restructure
Fixed import paths in test files and library files that were broken
after moving functions from nested directory structure.

Changed all imports from '../../../../../src/lib/' to '../../../src/lib/'
to account for the two-level reduction in directory nesting.

## Files Fixed
- netlify/functions/lib/error-handler.ts
- netlify/functions/lib/repository-validation.mts
- netlify/functions/__tests__/repository-validation.test.ts

## Testing
- ✅ api-suggest-reviewers test now passes (13 tests)
- ✅ repository-validation test now passes (18 tests)
